% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R

\name{chdage}
\alias{chdage}
\docType{data}
\title{
Coronary Heart Disease Study
}
\description{
A dataset containing the age and other attributes of almost 100 subjects selected to participate in a study. The variables are as follows:
}
\usage{chdage}

\format{
  A data frame with 100 observations on the following 3 variables.
  \describe{
    \item{\code{ID}}{identification code}
    \item{\code{AGE}}{age in years}
    \item{\code{CHD}}{presence (1) or absence (0) of evidence of significant coronary heart disease}
  }
}

\references{
[1] Hosmer, D. (2013). Wiley Series in Probability and Statistics Ser. : Applied Logistic Regression (3). New York: John Wiley & Sons, Incorporated.
}
\examples{
data(chdege)
## maybe str(chdege) ; plot(chdege) ...
}

\keyword{datasets}
