% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R

\name{icu}
\alias{icu}
\docType{data}
\title{
The icu Study.
}
\description{
The \code{icu} study data set consists of a sample of 200 subjects who were part of a much larger study on survival of patients following admission to an adult intensive care unit (\code{icu}). The variables are as follows:
}
\usage{icu}
\format{
  A data frame with 200 observations on the following 21 variables.
  \describe{
    \item{\code{ID}}{a numeric vector}
    \item{\code{STA}}{a numeric vector}
    \item{\code{AGE}}{a numeric vector}
    \item{\code{GENDER}}{a numeric vector}
    \item{\code{RACE}}{a numeric vector}
    \item{\code{SER}}{a numeric vector}
    \item{\code{CAN}}{a numeric vector}
    \item{\code{CRN}}{a numeric vector}
    \item{\code{INF}}{a numeric vector}
    \item{\code{CPR}}{a numeric vector}
    \item{\code{SYS}}{a numeric vector}
    \item{\code{HRA}}{a numeric vector}
    \item{\code{PRE}}{a numeric vector}
    \item{\code{TYP}}{a numeric vector}
    \item{\code{FRA}}{a numeric vector}
    \item{\code{PO2}}{a numeric vector}
    \item{\code{PH}}{a numeric vector}
    \item{\code{PCO}}{a numeric vector}
    \item{\code{BIC}}{a numeric vector}
    \item{\code{CRE}}{a numeric vector}
    \item{\code{LOC}}{a numeric vector}
  }
}
\references{
[1] Hosmer, D. (2013). Wiley Series in Probability and Statistics Ser. : Applied Logistic Regression (3). New York: John Wiley & Sons, Incorporated.
}
\examples{
data(icu)
## maybe str(icu) ; plot(icu) ...
}
\keyword{datasets}
