% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R

\name{pros}
\alias{pros}
\docType{data}
\title{
The Prostate Cancer Study
}
\description{
A third data set involves a study of patiens with cancer of the prostate.
}
\usage{pros}

\format{
  A data frame with 380 observations on the following 9 variables.
  \describe{
    \item{\code{ID}}{a numeric vector}
    \item{\code{CAPSULE}}{a numeric vector}
    \item{\code{AGE}}{a numeric vector}
    \item{\code{RACE}}{a numeric vector}
    \item{\code{DPROS}}{a numeric vector}
    \item{\code{DCAPS}}{a numeric vector}
    \item{\code{PSA}}{a numeric vector}
    \item{\code{VOL}}{a numeric vector}
    \item{\code{GLEASON}}{a numeric vector}
  }
}
\references{
[1] Hosmer, D. (2013). Wiley Series in Probability and Statistics Ser. : Applied Logistic Regression (3). New York: John Wiley & Sons, Incorporated.
}
\examples{
data(pros)
## maybe str(pros) ; plot(pros) ...
}
\keyword{datasets}
