\name{sampleTheta}
\alias{sampleTheta}
\title{Method for sampling theta values for the latent set model}
\usage{sampleTheta(y, dims, params, priors, verbose=FALSE, zcode=0, mh.scale=1)
}
\value{list of parameter values: w, z, and the updated theta}
\author{Chris DuBois}
\arguments{\item{y}{}
\item{dims}{}
\item{params}{}
\item{priors}{}
\item{verbose}{}
\item{zcode}{}
\item{mh.scale}{standard deviation of the Gaussian used in the Metropolis-Hastings step.}
}

