% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R

\name{uis}
\alias{uis}
\docType{data}
\title{
uis
}
\description{
uis
}
\usage{uis}
\format{
  A data frame with 575 observations on the following 9 variables.
  \describe{
    \item{\code{ID}}{a numeric vector}
    \item{\code{AGE}}{a numeric vector}
    \item{\code{BECK}}{a numeric vector}
    \item{\code{IVHX}}{a numeric vector}
    \item{\code{NDRUGTX}}{a numeric vector}
    \item{\code{RACE}}{a numeric vector}
    \item{\code{TREAT}}{a numeric vector}
    \item{\code{SITE}}{a numeric vector}
    \item{\code{DFREE}}{a numeric vector}
  }
}
\references{
[1] Hosmer, D. (2013). Wiley Series in Probability and Statistics Ser. : Applied Logistic Regression (3). New York: John Wiley & Sons, Incorporated.
}
\examples{
data(uis)
## maybe str(uis) ; plot(uis) ...
}
\keyword{uis}
