% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chdage.R
\name{chdage}
\alias{chdage}
\title{Coronary Heart Disease Study}
\format{
A data frame with 100 observations on the following 3 variables.
\describe{
 \item{\code{ID}}{identification code}
 \item{\code{AGE}}{age in years}
 \item{\code{CHD}}{presence (1) or absence (0) of evidence of significant coronary heart disease}
}
}
\usage{
chdage
}
\description{
Coronary Heart Disease Study
}
\examples{
  # data(chdage)
  # maybe str(chdage) ; plot(chdage) ...
}
\references{
Hosmer, D. (2013). Wiley Series in Probability and Statistics Ser. : Applied Logistic Regression (3). New York: John Wiley & Sons, Incorporated.
}
\keyword{cedage}
