% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.lsm.R
\name{confint.lsm}
\alias{confint.lsm}
\title{Confidence Intervals for \code{lsm}  Objects}
\usage{
\method{confint}{lsm}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{The type of prediction required. The default is on the scale of the linear predictors. 
The alternative \code{response} gives the predicted probabilities.}

\item{parm}{further arguments passed to or from other methods.}

\item{level}{The type of prediction required. The default is on the scale of the linear predictors. 
The alternative \code{response} gives the predicted probabilities.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
\code{lsm} returns an object of class "\code{lsm}".

 An object of class "\code{lsm}" is a list containing at least the
 following components:
 
\item{object}{a lsm object}
\item{parm}{parameter}
\item{level }{confidence levels}
\item{\dots}{additional parameters}
}
\description{
Provides a confint method for \code{lsm}  objects.
}
\details{
confint Method for lsm

The saturated model is characterized by the assumptions 1 and 2 presented in section 2.3 by Llinas (2006, ISSN:2389-8976).
}
\examples{
 #Hosmer, D. (2013) page 3: Age and coranary Heart Disease (CHD) Status of 20 subjects:
 #AGE <- c(20, 23, 24, 25, 25, 26, 26, 28, 28, 29, 30, 30, 30, 30, 30, 30, 30, 32, 33, 33)
 #CHD <- c(0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0)
 # data <- data.frame (CHD, AGE)
 # Ela <- lsm(CHD ~ AGE, family = binomial, data)
 # summary(Ela)
}
\references{
[1] Humberto Jesus Llinas. (2006). Accuracies in the theory of the logistic models. Revista Colombiana De Estadistica,29(2), 242-244.

[2] Hosmer, D. (2013). Wiley Series in Probability and Statistics Ser. : Applied Logistic Regression (3). New York: John Wiley & Sons, Incorporated.

[3] Chambers, J. M. and Hastie, T. J. (1992) Statistical Models in S. Wadsworth & Brooks/Cole.
}
\author{
Jorge Villalba Acevedo [cre, aut], Cartagena-Colombia.
}
