% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.lsm.R
\name{predict.lsm}
\alias{predict.lsm}
\title{Predict  Method for \code{lsm}  Objects}
\usage{
\method{predict}{lsm}(
  object,
  newdata,
  type = c("link", "response", "odd"),
  interval = c("none", "confidence", "prediction", "odd"),
  level = 0.95,
  ...
)
}
\arguments{
\item{object}{A fitted object of class \code{lsm}.}

\item{newdata}{Optionally, a data frame in which to look for variables with which to predict. 
If omitted, the fitted linear predictors are used.}

\item{type}{The type of prediction required. The default is on the scale of the linear predictors. 
The alternative \code{response} gives the predicted probabilities.}

\item{interval}{gives the}

\item{level}{gives the}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A vector or matrix of predictions.
 following components:
}
\description{
Obtains predictions from a fitted \code{lsm} object.
}
\details{
Predict Method for lsm Fits
}
