% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.lsm.R

\name{summary.lsm}
\alias{summary.lsm}
\title{
\code{summary} Method for \code{lsm} Objects
}

\description{
Provides a summary method for \code{lsm}  objects.
}

\usage{
\method{summary}{lsm}(object, ...)
}

\arguments{
  \item{object}{a lsm object}
   \item{\dots}{additional parameters}
}

\value{
Side effect: a summary table with results from lsm analysis.
}

\examples{
# Hosmer, D. (2013) page 3: Age and coranary Heart Disease (CHD) Status of 20 subjects:

 AGE <- c(20, 23, 24, 25, 25, 26, 26, 28, 28, 29, 30, 30, 30, 30, 30, 30, 30, 32, 33, 33)
 CHD <- c(0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0)
 data <- data.frame (CHD, AGE)
 Ela <- lsm(CHD ~ AGE, family = binomial, data)
 summary(Ela)
}

\references{
[1] Humberto Jesus Llinas. (2006). Accuracies in the theory of the logistic models. Revista Colombiana De Estadistica,29(2), 242-244.

[2] Hosmer, D. (2013). Wiley Series in Probability and Statistics Ser. : Applied Logistic Regression (3). New York: John Wiley & Sons, Incorporated.
}

\author{Jorge Villalba Acevedo}

\keyword{ summary.lsm }


 

